Moved grid allocation onto the MapGridComponent.
